requirements.txt
fastapi
uvicorn
replicate
python-dotenv
python-multipart
pillow